Added flow on site screen, navigation to individual site screen

+++ app/src/main/java/com/example/quickcast/room_db/dao/MessageDao.kt
@@ -18,5 +18,5 @@ interface MessageDao {
         FROM message_table
         WHERE siteId = :siteId
     """)
-    fun getMessagesDTOForSite(siteId: Int): Flow<List<MessageDTO>>
+    fun getMessagesDTOForSite(siteId: Long): Flow<List<MessageDTO>>
 }
